requirements.txt
# Web framework
fastapi>=0.104,<1.0
uvicorn[standard]>=0.24,<1.0
starlette>=0.27,<1.0
python-multipart>=0.0.6
pydantic-settings>=2.1,<3.0

# ML / Computer Vision
transformers>=4.36,<5.0
torch>=2.1,<3.0
opencv-python>=4.8,<5.0
pillow>=10.0,<12.0
numpy>=1.24,<3.0

# Security
cryptography>=41.0,<45.0

# Testing
pytest>=7.4,<9.0
pytest-cov>=4.1,<6.0
pytest-asyncio>=0.23,<1.0
httpx>=0.25,<1.0
